% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_post.R
\name{send_post}
\alias{send_post}
\title{Send Post}
\usage{
send_post(
  .letter = "Hello world!",
  .rider = NULL,
  .channel = NULL,
  .webhook = NULL
)
}
\arguments{
\item{.letter}{a string, default: \code{"Hello world!"}; should be enclosed
in quotes and contain the user's message}

\item{.rider}{a string, default: \code{NULL}; should be enclosed in quotes.
If left as NULL will fill in as \code{Sys.info()[["user"]]}}

\item{.channel}{a string, default: \code{NULL}; should be enclosed in quotes
and contain the name of a valid environment variable containing a channel
webhook string the user/.rider has access to}

\item{.webhook}{a string, default: \code{NULL}; recommend storing webhook
as a user environment variable or some other secure manner and extracting
as a character string to this argument}
}
\description{
Send a message to a Slack channel using R and webhooks.
}
\examples{
send_post(.letter = "foo", .channel = "bar")
}
\keyword{message}
\keyword{post}
\keyword{send}
\keyword{slack}
